menerapkan modularization dengan memindahkan file file yang ada di core ke modul core yang telah dibuat. kemudian menambahkan dependency yang ditbutuhkan pada build.gradle core. kemudian pada modul utama mengganti pemanggilan iimport sesaui dengan import modul core
