+++ app/src/main/java/com/dicoding/tourismapp/MyApplication.kt
@@ -1,11 +1,11 @@
 package com.dicoding.tourismapp
 
 import android.app.Application
-import com.dicoding.tourismapp.core.di.databaseModule
-import com.dicoding.tourismapp.core.di.networkModule
-import com.dicoding.tourismapp.core.di.repositoryModule
 import com.dicoding.tourismapp.di.useCaseModule
 import com.dicoding.tourismapp.di.viewModelModule
+import com.example.core.di.databaseModule
+import com.example.core.di.networkModule
+import com.example.core.di.repositoryModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
