--- app/src/main/java/com/dicoding/tourismapp/core/data/Resource.kt
@@ -1,7 +1,0 @@
-package com.dicoding.tourismapp.core.data
-
-sealed class Resource<T>(val data: T? = null, val message: String? = null) {
-    class Success<T>(data: T) : Resource<T>(data)
-    class Loading<T>(data: T? = null) : Resource<T>(data)
-    class Error<T>(message: String, data: T? = null) : Resource<T>(data, message)
-}
