--- app/src/main/java/com/dicoding/tourismapp/core/data/TourismRepository.kt
@@ -1,74 +1,0 @@
-package com.dicoding.tourismapp.core.data
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.map
-import com.dicoding.tourismapp.core.data.source.remote.network.ApiResponse
-import com.dicoding.tourismapp.core.data.source.local.LocalDataSource
-import com.dicoding.tourismapp.core.data.source.local.entity.TourismEntity
-import com.dicoding.tourismapp.core.data.source.remote.RemoteDataSource
-import com.dicoding.tourismapp.core.data.source.remote.response.PlacesItem
-import com.dicoding.tourismapp.core.domain.model.Tourism
-import com.dicoding.tourismapp.core.domain.repository.ITourismRepository
-import com.dicoding.tourismapp.core.utils.AppExecutors
-import com.dicoding.tourismapp.core.utils.DataMapper
-import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
-import io.reactivex.rxjava3.core.Flowable
-import io.reactivex.rxjava3.schedulers.Schedulers
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-
-class TourismRepository(
-    private val remoteDataSource: RemoteDataSource,
-    private val localDataSource: LocalDataSource,
-    private val appExecutors: AppExecutors
-) : ITourismRepository {
-
-    companion object {
-        @Volatile
-        private var instance: TourismRepository? = null
-
-        // dihapus karena sudah menggunakan koin
-//        fun getInstance(
-//            remoteData: RemoteDataSource,
-//            localData: LocalDataSource,
-//            appExecutors: AppExecutors
-//        ): TourismRepository =
-//            instance ?: synchronized(this) {
-//                instance ?: TourismRepository(remoteData, localData, appExecutors)
-//            }
-    }
-
-    // mengubah tipe data pada TourismRepository yang sebelumnya menggunakan TourismEntity menjadi Tourism
-
-    // selanjutnya sesuaikan disini menjadi Flowable
-    override fun getAllTourism(): Flow<Resource<List<Tourism>>> =
-        object : NetworkBoundResource<List<Tourism>, List<PlacesItem>>() {
-            override fun loadFromDB(): Flow<List<Tourism>> {
-                return localDataSource.getAllTourism().map { DataMapper.mapEntitiesToDomain(it) }
-            }
-
-            override fun shouldFetch(data: List<Tourism>?): Boolean =
-                data.isNullOrEmpty() // mengambil data dari internet hanya jika data di database kosong
-//                 true // ganti dengan true jika ingin selalu mengambil data dari internet
-
-            override suspend fun createCall(): Flow<ApiResponse<List<PlacesItem>>> =
-                remoteDataSource.getAllTourism()
-
-            override suspend fun saveCallResult(data: List<PlacesItem>) {
-                val tourismList = DataMapper.mapResponsesToEntities(data)
-                localDataSource.insertTourism(tourismList)
-            }
-        }.asFlow()
-
-    override fun getFavoriteTourism(): Flow<List<Tourism>> {
-        return localDataSource.getFavoriteTourism().map {
-            DataMapper.mapEntitiesToDomain(it)
-        }
-    }
-
-    override fun setFavoriteTourism(tourism: Tourism, state: Boolean) {
-        val tourismEntity = DataMapper.mapDomainToEntity(tourism)
-        appExecutors.diskIO().execute { localDataSource.setFavoriteTourism(tourismEntity, state) }
-    }
-}
-
