--- app/src/main/java/com/dicoding/tourismapp/core/data/source/local/room/TourismDao.kt
@@ -1,24 +1,0 @@
-package com.dicoding.tourismapp.core.data.source.local.room
-
-import androidx.lifecycle.LiveData
-import androidx.room.*
-import com.dicoding.tourismapp.core.data.source.local.entity.TourismEntity
-import io.reactivex.rxjava3.core.Completable
-import io.reactivex.rxjava3.core.Flowable
-import kotlinx.coroutines.flow.Flow
-
-@Dao
-interface TourismDao {
-
-    @Query("SELECT * FROM tourism")
-    fun getAllTourism(): Flow<List<TourismEntity>>
-
-    @Query("SELECT * FROM tourism where isFavorite = 1")
-    fun getFavoriteTourism(): Flow<List<TourismEntity>>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertTourism(tourism: List<TourismEntity>) // menggunakan suspend supaya bisa dilakukan secara asynchronous
-
-    @Update
-    fun updateFavoriteTourism(tourism: TourismEntity)
-}
