--- app/src/main/java/com/dicoding/tourismapp/core/data/source/remote/RemoteDataSource.kt
@@ -1,60 +1,0 @@
-package com.dicoding.tourismapp.core.data.source.remote
-
-import android.annotation.SuppressLint
-import android.os.Handler
-import android.os.Looper
-import android.util.Log
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import com.dicoding.tourismapp.core.data.source.remote.network.ApiResponse
-import com.dicoding.tourismapp.core.data.source.remote.network.ApiService
-import com.dicoding.tourismapp.core.data.source.remote.response.PlacesItem
-import com.dicoding.tourismapp.core.data.source.remote.response.TourismResponse
-import com.dicoding.tourismapp.core.utils.JsonHelper
-import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
-import io.reactivex.rxjava3.core.BackpressureStrategy
-import io.reactivex.rxjava3.core.Flowable
-import io.reactivex.rxjava3.schedulers.Schedulers
-import io.reactivex.rxjava3.subjects.PublishSubject
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOn
-import org.json.JSONException
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-
-class RemoteDataSource(private val apiService: ApiService) {
-    companion object {
-        @Volatile
-        private var instance: RemoteDataSource? = null
-
-        // dihapus karena sudah menggunakan koin
-//        fun getInstance(service: ApiService): RemoteDataSource =
-//            instance ?: synchronized(this) {
-//                instance ?: RemoteDataSource(service)
-//            }
-    }
-
-    // mengubah LiveData menjadi Flowable
-    @SuppressLint("CheckResult")
-    suspend fun getAllTourism(): Flow<ApiResponse<List<PlacesItem>>>  {
-        return flow {
-            try {
-                val response = apiService.getList()
-                val dataArray = response.places
-                if (dataArray.isNotEmpty()){
-                    emit(ApiResponse.Success(response.places))
-                } else {
-                    emit(ApiResponse.Empty)
-                }
-            } catch (e : Exception) {
-                emit(ApiResponse.Error(e.toString()))
-                Log.e("RemoteDataSource", e.toString())
-            }
-        }.flowOn(Dispatchers.IO)
-
-    }
-}
-
