--- app/src/main/java/com/dicoding/tourismapp/core/data/source/remote/network/ApiService.kt
@@ -1,11 +1,0 @@
-package com.dicoding.tourismapp.core.data.source.remote.network
-
-import com.dicoding.tourismapp.core.data.source.remote.response.TourismResponse
-import io.reactivex.rxjava3.core.Flowable
-import retrofit2.Call
-import retrofit2.http.GET
-
-interface ApiService {
-    @GET("list")
-    suspend fun getList(): TourismResponse // hapus call, tambahkan suspend
-}
