--- app/src/main/java/com/dicoding/tourismapp/core/di/CoreModule.kt
@@ -1,54 +1,0 @@
-package com.dicoding.tourismapp.core.di
-
-import androidx.room.Room
-import com.dicoding.tourismapp.core.data.TourismRepository
-import com.dicoding.tourismapp.core.data.source.local.LocalDataSource
-import com.dicoding.tourismapp.core.data.source.local.room.TourismDatabase
-import com.dicoding.tourismapp.core.data.source.remote.RemoteDataSource
-import com.dicoding.tourismapp.core.data.source.remote.network.ApiService
-import com.dicoding.tourismapp.core.domain.repository.ITourismRepository
-import com.dicoding.tourismapp.core.utils.AppExecutors
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import org.koin.android.ext.koin.androidContext
-import org.koin.dsl.module
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-val databaseModule = module {
-    factory {
-        get<TourismDatabase>().tourismDao()
-    }
-    single {
-        Room.databaseBuilder(
-            androidContext(),
-            TourismDatabase::class.java, "Tourism.db"
-        ).fallbackToDestructiveMigration().build()
-    }
-}
-
-val networkModule = module {
-    single {
-        OkHttpClient.Builder()
-            .addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-            .connectTimeout(120, TimeUnit.SECONDS)
-            .readTimeout(120, TimeUnit.SECONDS)
-            .build()
-    }
-    single {
-        val retrofit = Retrofit.Builder()
-            .baseUrl("https://tourism-api.dicoding.dev/")
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(get())
-            .build()
-        retrofit.create(ApiService::class.java)
-    }
-}
-
-val repositoryModule = module {
-    single { LocalDataSource(get()) }
-    single { RemoteDataSource(get()) }
-    factory { AppExecutors() }
-    single<ITourismRepository> { TourismRepository(get(), get(), get()) }
-}
