--- app/src/main/java/com/dicoding/tourismapp/core/domain/usecase/TourismInteractor.kt
@@ -1,16 +1,0 @@
-package com.dicoding.tourismapp.core.domain.usecase
-
-import androidx.lifecycle.LiveData
-import com.dicoding.tourismapp.core.data.Resource
-import com.dicoding.tourismapp.core.domain.model.Tourism
-import com.dicoding.tourismapp.core.domain.repository.ITourismRepository
-import io.reactivex.rxjava3.core.Flowable
-
-class TourismInteractor(private val tourismRepository: ITourismRepository): TourismUseCase {
-    override fun getAllTourism() = tourismRepository.getAllTourism()
-
-    override fun getFavoriteTourism() = tourismRepository.getFavoriteTourism()
-
-    override fun setFavoriteTourism(tourism: Tourism, state: Boolean) = tourismRepository.setFavoriteTourism(tourism, state)
-
-}
