--- app/src/main/java/com/dicoding/tourismapp/core/domain/usecase/TourismUseCase.kt
@@ -1,18 +1,0 @@
-package com.dicoding.tourismapp.core.domain.usecase
-
-import androidx.lifecycle.LiveData
-import com.dicoding.tourismapp.core.data.Resource
-import com.dicoding.tourismapp.core.domain.model.Tourism
-import io.reactivex.rxjava3.core.Flowable
-import kotlinx.coroutines.flow.Flow
-
-interface TourismUseCase {
-
-    // mengubah menjadi flowable
-    fun getAllTourism(): Flow<Resource<List<Tourism>>>
-
-    fun getFavoriteTourism(): Flow<List<Tourism>>
-
-    fun setFavoriteTourism(tourism: Tourism, state: Boolean)
-
-}
