+++ app/src/main/java/com/dicoding/tourismapp/di/AppModule.kt
@@ -1,10 +1,10 @@
 package com.dicoding.tourismapp.di
 
-import com.dicoding.tourismapp.core.domain.usecase.TourismInteractor
-import com.dicoding.tourismapp.core.domain.usecase.TourismUseCase
 import com.dicoding.tourismapp.detail.DetailTourismViewModel
 import com.dicoding.tourismapp.favorite.FavoriteViewModel
 import com.dicoding.tourismapp.home.HomeViewModel
+import com.example.core.domain.usecase.TourismInteractor
+import com.example.core.domain.usecase.TourismUseCase
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
