+++ app/src/main/java/com/dicoding/tourismapp/home/HomeViewModel.kt
@@ -2,9 +2,7 @@ package com.dicoding.tourismapp.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
-import androidx.lifecycle.toLiveData
-import com.dicoding.tourismapp.core.data.TourismRepository
-import com.dicoding.tourismapp.core.domain.usecase.TourismUseCase
+import com.example.core.domain.usecase.TourismUseCase
 
 class HomeViewModel(tourismUseCase: TourismUseCase) : ViewModel() {
