+++ core/src/main/java/com/example/core/data/Resource.kt
@@ -1,0 +1,7 @@
+package com.example.core.data
+
+sealed class Resource<T>(val data: T? = null, val message: String? = null) {
+    class Success<T>(data: T) : com.example.core.data.Resource<T>(data)
+    class Loading<T>(data: T? = null) : com.example.core.data.Resource<T>(data)
+    class Error<T>(message: String, data: T? = null) : com.example.core.data.Resource<T>(data, message)
+}
