+++ core/src/main/java/com/example/core/data/TourismRepository.kt
@@ -1,0 +1,69 @@
+package com.example.core.data
+
+
+import com.example.core.data.source.local.LocalDataSource
+import com.example.core.data.source.remote.RemoteDataSource
+import com.example.core.data.source.remote.network.ApiResponse
+import com.example.core.data.source.remote.response.PlacesItem
+import com.example.core.domain.model.Tourism
+import com.example.core.domain.repository.ITourismRepository
+import com.example.core.utils.AppExecutors
+import com.example.core.utils.DataMapper
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+
+class TourismRepository(
+    private val remoteDataSource: RemoteDataSource,
+    private val localDataSource: LocalDataSource,
+    private val appExecutors: AppExecutors
+) : ITourismRepository {
+
+    companion object {
+        @Volatile
+        private var instance: TourismRepository? = null
+
+        // dihapus karena sudah menggunakan koin
+//        fun getInstance(
+//            remoteData: RemoteDataSource,
+//            localData: LocalDataSource,
+//            appExecutors: AppExecutors
+//        ): TourismRepository =
+//            instance ?: synchronized(this) {
+//                instance ?: TourismRepository(remoteData, localData, appExecutors)
+//            }
+    }
+
+    // mengubah tipe data pada TourismRepository yang sebelumnya menggunakan TourismEntity menjadi Tourism
+
+    // selanjutnya sesuaikan disini menjadi Flowable
+    override fun getAllTourism(): Flow<com.example.core.data.Resource<List<Tourism>>> =
+        object : com.example.core.data.NetworkBoundResource<List<Tourism>, List<PlacesItem>>() {
+            override fun loadFromDB(): Flow<List<Tourism>> {
+                return localDataSource.getAllTourism().map { DataMapper.mapEntitiesToDomain(it) }
+            }
+
+            override fun shouldFetch(data: List<Tourism>?): Boolean =
+                data.isNullOrEmpty() // mengambil data dari internet hanya jika data di database kosong
+//                 true // ganti dengan true jika ingin selalu mengambil data dari internet
+
+            override suspend fun createCall(): Flow<ApiResponse<List<PlacesItem>>> =
+                remoteDataSource.getAllTourism()
+
+            override suspend fun saveCallResult(data: List<PlacesItem>) {
+                val tourismList = DataMapper.mapResponsesToEntities(data)
+                localDataSource.insertTourism(tourismList)
+            }
+        }.asFlow()
+
+    override fun getFavoriteTourism(): Flow<List<Tourism>> {
+        return localDataSource.getFavoriteTourism().map {
+            DataMapper.mapEntitiesToDomain(it)
+        }
+    }
+
+    override fun setFavoriteTourism(tourism: Tourism, state: Boolean) {
+        val tourismEntity = DataMapper.mapDomainToEntity(tourism)
+        appExecutors.diskIO().execute { localDataSource.setFavoriteTourism(tourismEntity, state) }
+    }
+}
+
