+++ core/src/main/java/com/example/core/data/source/local/LocalDataSource.kt
@@ -1,0 +1,30 @@
+package com.example.core.data.source.local
+
+import com.example.core.data.source.local.entity.TourismEntity
+import com.example.core.data.source.local.room.TourismDao
+import kotlinx.coroutines.flow.Flow
+
+class LocalDataSource(private val tourismDao: TourismDao) {
+
+    companion object {
+        private var instance: LocalDataSource? = null
+
+        // dihapus karena sudah menggunakan koin
+//        fun getInstance(tourismDao: TourismDao): LocalDataSource =
+//            instance ?: synchronized(this) {
+//                instance ?: LocalDataSource(tourismDao)
+//            }
+    }
+
+
+    fun getAllTourism(): Flow<List<TourismEntity>> = tourismDao.getAllTourism()
+
+    fun getFavoriteTourism(): Flow<List<TourismEntity>> = tourismDao.getFavoriteTourism()
+
+    suspend fun insertTourism(tourismList: List<TourismEntity>) = tourismDao.insertTourism(tourismList)
+
+    fun setFavoriteTourism(tourism: TourismEntity, newState: Boolean) {
+        tourism.isFavorite = newState
+        tourismDao.updateFavoriteTourism(tourism)
+    }
+}
