+++ core/src/main/java/com/example/core/data/source/local/room/TourismDao.kt
@@ -1,0 +1,22 @@
+package com.example.core.data.source.local.room
+
+import androidx.lifecycle.LiveData
+import androidx.room.*
+import com.example.core.data.source.local.entity.TourismEntity
+import kotlinx.coroutines.flow.Flow
+
+@Dao
+interface TourismDao {
+
+    @Query("SELECT * FROM tourism")
+    fun getAllTourism(): Flow<List<TourismEntity>>
+
+    @Query("SELECT * FROM tourism where isFavorite = 1")
+    fun getFavoriteTourism(): Flow<List<TourismEntity>>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertTourism(tourism: List<TourismEntity>) // menggunakan suspend supaya bisa dilakukan secara asynchronous
+
+    @Update
+    fun updateFavoriteTourism(tourism: TourismEntity)
+}
