+++ core/src/main/java/com/example/core/data/source/local/room/TourismDatabase.kt
@@ -1,0 +1,31 @@
+package com.example.core.data.source.local.room
+
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import com.example.core.data.source.local.entity.TourismEntity
+
+@Database(entities = [TourismEntity::class], version = 1, exportSchema = false)
+abstract class TourismDatabase : RoomDatabase() {
+
+    abstract fun tourismDao(): TourismDao
+
+    companion object {
+        @Volatile
+        private var INSTANCE: TourismDatabase? = null
+
+        // dihapus karena sudah menggunakan koin
+//        fun getInstance(context: Context): TourismDatabase =
+//            INSTANCE ?: synchronized(this) {
+//            val instance = Room.databaseBuilder(
+//                context.applicationContext,
+//                TourismDatabase::class.java,
+//                "Tourism.db"
+//            )
+//                .fallbackToDestructiveMigration()
+//                .build()
+//            INSTANCE = instance
+//            instance
+//        }
+    }
+}
