+++ core/src/main/java/com/example/core/data/source/remote/RemoteDataSource.kt
@@ -1,0 +1,45 @@
+package com.example.core.data.source.remote
+
+import android.annotation.SuppressLint
+import android.util.Log
+import com.example.core.data.source.remote.network.ApiResponse
+import com.example.core.data.source.remote.network.ApiService
+import com.example.core.data.source.remote.response.PlacesItem
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+
+class RemoteDataSource(private val apiService: ApiService) {
+    companion object {
+        @Volatile
+        private var instance: RemoteDataSource? = null
+
+        // dihapus karena sudah menggunakan koin
+//        fun getInstance(service: ApiService): RemoteDataSource =
+//            instance ?: synchronized(this) {
+//                instance ?: RemoteDataSource(service)
+//            }
+    }
+
+    // mengubah LiveData menjadi Flowable
+    @SuppressLint("CheckResult")
+    suspend fun getAllTourism(): Flow<ApiResponse<List<PlacesItem>>>  {
+        return flow {
+            try {
+                val response = apiService.getList()
+                val dataArray = response.places
+                if (dataArray.isNotEmpty()){
+                    emit(ApiResponse.Success(response.places))
+                } else {
+                    emit(ApiResponse.Empty)
+                }
+            } catch (e : Exception) {
+                emit(ApiResponse.Error(e.toString()))
+                Log.e("RemoteDataSource", e.toString())
+            }
+        }.flowOn(Dispatchers.IO)
+
+    }
+}
+
