+++ core/src/main/java/com/example/core/data/source/remote/network/ApiConfig.kt
@@ -1,0 +1,27 @@
+package com.example.core.data.source.remote.network
+
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+
+object ApiConfig {
+    // fungsi untuk menampilkan hasil response dari API ke dalam LogCat
+    private fun provideOkHttpClient(): OkHttpClient {
+        return OkHttpClient.Builder()
+            .addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
+            .connectTimeout(120, TimeUnit.SECONDS)
+            .readTimeout(120, TimeUnit.SECONDS)
+            .build()
+    }
+
+    fun provideApiService(): ApiService {
+        val retrofit = Retrofit.Builder()
+            .baseUrl("https://tourism-api.dicoding.dev/")
+            .addConverterFactory(GsonConverterFactory.create())
+            .client(provideOkHttpClient())
+            .build()
+        return retrofit.create(ApiService::class.java)
+    }
+}
