+++ core/src/main/java/com/example/core/data/source/remote/network/ApiResponse.kt
@@ -1,0 +1,7 @@
+package com.example.core.data.source.remote.network
+
+sealed class ApiResponse<out R> {
+    data class Success<out T>(val data: T) : ApiResponse<T>()
+    data class Error(val errorMessage: String) : ApiResponse<Nothing>()
+    data object Empty : ApiResponse<Nothing>()
+}
