+++ core/src/main/java/com/example/core/utils/DataMapper.kt
@@ -1,0 +1,55 @@
+package com.example.core.utils
+
+import com.example.core.data.source.local.entity.TourismEntity
+import com.example.core.data.source.remote.response.PlacesItem
+import com.example.core.domain.model.Tourism
+
+object DataMapper {
+    fun mapResponsesToEntities(input: List<PlacesItem>): List<TourismEntity> {
+        val tourismList = ArrayList<TourismEntity>()
+        input.map {
+            val tourism = TourismEntity(
+                tourismId = it.id,
+                description = it.description,
+                name = it.name,
+                address = it.address,
+                latitude = it.latitude,
+                longitude = it.longitude,
+                like = it.like,
+                image = it.image,
+                isFavorite = false
+            )
+            tourismList.add(tourism)
+        }
+        return tourismList
+    }
+
+    // mengubah kelas TourismEntity menjadi model Tourism (Domain)
+    fun mapEntitiesToDomain(input: List<TourismEntity>): List<Tourism> =
+        input.map {
+            Tourism(
+                tourismId = it.tourismId,
+                description = it.description,
+                name = it.name,
+                address = it.address,
+                latitude = it.latitude,
+                longitude = it.longitude,
+                like = it.like,
+                image = it.image,
+                isFavorite = it.isFavorite
+            )
+        }
+
+    // mengubah model yg di domain menjadi model yang di entity
+    fun mapDomainToEntity(input: Tourism) = TourismEntity(
+        tourismId = input.tourismId,
+        description = input.description,
+        name = input.name,
+        address = input.address,
+        latitude = input.latitude,
+        longitude = input.longitude,
+        like = input.like,
+        image = input.image,
+        isFavorite = input.isFavorite
+    )
+}
