+++ settings.gradle.kts
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Tourism App"
 include(":app")
+include(":core")
